FLAS-40-Cards refactor: - addCard modal form
