--- src/features/cards/createCard/addCardForm/addItemForm/index.ts
@@ -1,1 +1,0 @@
-export * from './addItemForm'
