--- src/features/cards/createCard/addCardForm/index.ts
@@ -1,1 +1,0 @@
-export * from './addCardForm'
