+++ src/features/cards/createCard/createCard.tsx
@@ -1,13 +1,11 @@
 import { Modal } from '@/common/ui/modal'
 import { Card, useCreateCardMutation } from '@/services/cards'
 
-import { AddCardForm } from './addCardForm/addCardForm'
+import { CardForm } from '../cardForm'
 
-/** Контейнерная компонента createCard для логики запросов */
 type Props = {
   card?: Card
   deckId: string
-  disabled?: boolean
   isOpen: boolean
   onOpenChange: (isOpen: boolean) => void
   title: string
@@ -22,7 +20,7 @@ export const CreateCard = ({ card, deckId, isOpen, onOpenChange, title }: Props)
 
   return (
     <Modal onOpenChange={onOpenChange} open={isOpen} title={title}>
-      <AddCardForm card={card} onCreateCard={onCreateCard} onOpenChange={onOpenChange} />
+      <CardForm card={card} onCreateCard={onCreateCard} onOpenChange={onOpenChange} />
     </Modal>
   )
 }
